began text
